Rat Death & Sound
Added textures for a male adult rat and updated AdultRat to conform with all changes to rat e.g now being a subclass of RenderObjects. Added SoundClip.java to allow for sound effects, sounds must be of WAV format and stored in the 'Sounds' folder. Updated the ratDeath method in AdultRat and BabyRat.